refactor: New CommonTheme created for define common settings for both Light and Dark themes.
