+++ styles/themes/LightTheme.ts
@@ -1,18 +1,9 @@
-import { Theme, createTheme, responsiveFontSizes } from "@mui/material/styles";
+import { Theme, createTheme } from "@mui/material/styles";
 import "@/styles/globals.css";
+import commonTheme from "@/styles/themes/CommonTheme";
 
-// Augment the palette to include new background colors
-declare module '@mui/material/styles' {
-  interface TypeBackground {
-    surface1: string;
-  }
-
-  interface TypeBackgroundOptions {
-    surface1?: string;
-  }
-}
-
-let lightTheme: Theme = createTheme({
+const lightTheme: Theme = createTheme({
+  ...commonTheme,
   palette: {
     mode: "light",
     primary: {
@@ -31,34 +22,6 @@ let lightTheme: Theme = createTheme({
       surface1: "#F8F8F8",
     },
   },
-  typography: {
-    fontFamily: '"Lato", "Helvetica", "Arial", sans-serif',
-    fontSize: 14,
-    h1: {
-      fontFamily: '"Montserrat", "Helvetica", "Arial", sans-serif',
-      fontWeight: 700,
-    },
-    h2: {
-      fontFamily: '"Montserrat", "Helvetica", "Arial", sans-serif',
-    },
-    h3: {
-      fontFamily: '"Montserrat", "Helvetica", "Arial", sans-serif',
-    },
-    h4: {
-      fontFamily: '"Montserrat", "Helvetica", "Arial", sans-serif',
-      fontWeight: "bold",
-    },
-    h5: {
-      fontFamily: '"Montserrat", "Helvetica", "Arial", sans-serif',
-      fontWeight: "bold",
-    },
-    h6: {
-      fontFamily: '"Montserrat", "Helvetica", "Arial", sans-serif',
-      fontWeight: "bold",
-    },
-  },
 });
-
-lightTheme = responsiveFontSizes(lightTheme);
 
 export default lightTheme;
